refactor: 사이드바 관련 컴포넌트 이동 (#80)
